+++ docs/reference/changelog.rst
@@ -7,10 +7,18 @@ Changelog
 5.2.1 (2025-XX-YY)
 ------------------
 
+Application
+===========
+
+- Fix an issue where the fetch-service would fail to find the network used
+  by LXD containers.
+
 Commands
 ========
 
 - Reduce spread verbosity level when running the ``test`` command.
+
+For a complete list of commits, check out the `5.2.1`_ release on GitHub.
 
 5.2.0 (2025-04-25)
 ------------------
@@ -758,3 +766,4 @@ For a complete list of commits, check out the `2.7.0`_ release on GitHub.
 .. _5.0.4: https://github.com/canonical/craft-application/releases/tag/5.0.4
 .. _5.1.0: https://github.com/canonical/craft-application/releases/tag/5.1.0
 .. _5.2.0: https://github.com/canonical/craft-application/releases/tag/5.2.0
+.. _5.2.1: https://github.com/canonical/craft-application/releases/tag/5.2.1